[WIP] Hooked up White Noise to preivew
I intend for the algorithms to be completely abstracted and separated. The idea is, that the settings panel displays the correct AlgorithmSettingsPanel component which displays settings specific to that algorithm and upon clicking the generate button, it uses these settings (which are read from member gui components of this panel) to call an algorithm  function. Basically, it's the settings panel that generates the image, and it relays it to preview label. this is very much subject to change, this is the design I came up with that might be adjusted once I hit an unexpected roadblock.