Somewhat working prototype
I'm not satisfied with current implementation, It seems sketchy, duct tape & gum style and I need to think of something better, hopefully without making my AlgorithmManager* a metatype to pass it to QVariant (those macros scare me). And I'm still not sure if having a GUI component responsible for generation at the same time is a good idea. fml