Added support for seed in WN and CA
Did some copy and paste, and it occured to me that it could be usefull to extract some classes and derive from them, since a lot of these managers have common features. I didn't do it because I do not plan to extend this application. If I will extend it someday, I will do this